refactor: move the preview watcher into a separate composable
